Fix 3.6 linux compile (#4166)
* std::sinf not work on linux, if not  c++17

+++ native/cocos/primitive/Box.cpp
@@ -74,7 +74,7 @@ IGeometry box(const cc::optional<IBoxOptions> &options) {
     std::vector<uint32_t> indices;
     const Vec3            minPos(-hw, -hh, -hl);
     const Vec3            maxPos(hw, hh, hl);
-    float                 boundingRadius{sqrt(hw * hw + hh * hh + hl * hl)};
+    float                 boundingRadius{static_cast<float>(sqrt(hw * hw + hh * hh + hl * hl))};
 
     auto buildPlane = [&](uint32_t side, uint32_t uSegments, uint32_t vSegments) {
         float                  u           = 0;
